refactor: separate adaptor module files and better naming.
