+++ pkg/adaptor/response_to_openai.go
@@ -1,0 +1,1 @@
+package adaptor
